Restructure sources; methods in same source as support classes
API docs build and work as before.

71 tests, 86 assertions, 0 failures, 0 errors, 0 skips
Coverage: 900 / 900 LOC (100.0%) covered.
RuboCop: 11 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 309.0; method average 4.5; max 9.5 (CryptIdent::SignIn#call)
Reek: 0 total warnings
Inch: Nothing to suggest
