+++ docs/file.CODE_OF_CONDUCT.html
@@ -135,7 +135,7 @@ available at <a href="http://contributor-covenant.org/version/1/4/">http://contr
 </div></div>
 
       <div id="footer">
-  Generated on Fri Jan 25 02:28:53 2019 by
+  Generated on Sat Jan 26 00:46:02 2019 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.16 (ruby-2.6.0).
 </div>
