+++ docs/file.README.html
@@ -678,7 +678,7 @@
 </div></div>
 
       <div id="footer">
-  Generated on Fri Jan 25 02:28:53 2019 by
+  Generated on Sat Jan 26 00:46:02 2019 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.16 (ruby-2.6.0).
 </div>
